Updated the example to use unicode characters. Added a test to cover this case. There was a parsing bug where I was evaluating the letter bytes in the string instead of operating on the runes for the Runes. Fixing this in the Trie builder fixed the example and the test.
Added character and byte counts to the outputs. Fixed bug in state_fifo where the write pointer was not getting reset properly. This was a non-deterministic bug, unrelated to multi-threading. Sometimes it worked, sometimes it didn't. Fixed by resetting the slice size back to zero, which I did not do.

While this was clearly a mistake I had made (the read pointer was reset, but new items kept getting added onto the back and we'd never reach them), there's something wrong in Go where this was non-deterministic.